move check_login to utils
