+++ ncoreparser/client.py
@@ -1,5 +1,4 @@
 import os
-import functools
 import httpx
 from ncoreparser.data import (
     URLs,
@@ -20,18 +19,8 @@ from ncoreparser.parser import (
     ActivityParser,
     RecommendedParser
 )
-from ncoreparser.util import Size
+from ncoreparser.util import Size, check_login
 from ncoreparser.torrent import Torrent
-
-
-def _check_login(func):
-    @functools.wraps(func)
-    def wrapper(self, *args, **kwargs):
-        if not self._logged_in:  # pylint: disable=protected-access
-            raise NcoreConnectionError("Cannot login to tracker. "
-                                       f"Please use {Client.login.__name__} function first.")
-        return func(self, *args, **kwargs)
-    return wrapper
 
 
 class Client:
@@ -60,7 +49,7 @@ class Client:
                                        f"credentials for user: '{username}'")
         self._logged_in = True
 
-    @_check_login
+    @check_login
     # pylint: disable=too-many-arguments, too-many-positional-arguments
     def search(
         self,
@@ -91,7 +80,7 @@ class Client:
             page_count += 1
         return torrents[:number]
 
-    @_check_login
+    @check_login
     def get_torrent(self, id, **ext_params):
         url = URLs.DETAIL_PATTERN.value.format(id=id)
         try:
@@ -103,7 +92,7 @@ class Client:
         params.update(ext_params)
         return Torrent(**params)
 
-    @_check_login
+    @check_login
     def get_by_rss(self, url):
         try:
             content = self._client.get(url)
@@ -115,7 +104,7 @@ class Client:
             torrents.append(self.get_torrent(id))
         return torrents
 
-    @_check_login
+    @check_login
     def get_by_activity(self):
         try:
             content = self._client.get(URLs.ACTIVITY.value)
@@ -135,7 +124,7 @@ class Client:
                                              rate=float(rate)))
         return torrents
 
-    @_check_login
+    @check_login
     def get_recommended(self, type=None):
         try:
             content = self._client.get(URLs.RECOMMENDED.value)
@@ -145,7 +134,7 @@ class Client:
         all_recommended = [self.get_torrent(id) for id in self._recommended_parser.get_ids(content.text)]
         return [torrent for torrent in all_recommended if not type or torrent['type'] == type]
 
-    @_check_login
+    @check_login
     def download(self, torrent, path, override=False):
         file_path, url = torrent.prepare_download(path)
         try:
