+++ ncoreparser/util.py
@@ -1,4 +1,6 @@
 import datetime
+import functools
+from ncoreparser.error import NcoreConnectionError
 
 
 class Size:
@@ -83,3 +85,13 @@ class Size:
 
 def parse_datetime(date, time):
     return datetime.datetime.strptime(f"{date}_{time}", "%Y-%m-%d_%H:%M:%S")
+
+
+def check_login(func):
+    @functools.wraps(func)
+    def wrapper(self, *args, **kwargs):
+        if not self._logged_in:  # pylint: disable=protected-access
+            raise NcoreConnectionError("Cannot login to tracker. "
+                                       f"Please use {self.login.__name__} function first.")
+        return func(self, *args, **kwargs)
+    return wrapper
